Collision Prevention IslandCircle
IslandCircle collision now works with the methods .getX() and .getY()
copied from Island.java and overwritten in IslandCircle.java so that it
takes the x and y from IslandCircle specifically and does not return
zero (null for in). I have tested it thoroughly and have tested it in
panelbranch and it visibly prevents collision of circles even when the
map is swarmed with islands but if it doesnt just revert. I'm tired.